test: comment out view tests, not useful in that state
We'll come up with new view unit tests in a separate PR.
