+++ packages/web-app-files/tests/unit/composables/upload/useUploadHelpers.spec.ts
@@ -1,5 +1,8 @@
 import { useUploadHelpers } from '../../../../src/composables/upload'
 import { createWrapper } from './spec'
+import { mockDeep } from 'jest-mock-extended'
+import { SpaceResource } from 'web-client/src/helpers'
+import { ComputedRef, computed } from '@vue/composition-api'
 
 describe('useUploadHelpers', () => {
   const currentPathMock = 'path'
@@ -9,35 +12,19 @@ describe('useUploadHelpers', () => {
     expect(useUploadHelpers).toBeDefined()
   })
 
-  it('returns the correct current path', () => {
-    createWrapper(
-      () => {
-        const { currentPath } = useUploadHelpers()
-        expect(currentPath.value).toBe(`${currentPathMock}/`)
-      },
-      { currentItem: currentPathMock, fileUrl: uploadPathMock }
-    )
-  })
-
-  it('returns the correct uploadPath', () => {
-    createWrapper(
-      () => {
-        const { uploadPath } = useUploadHelpers()
-        expect(uploadPath.value).toBe(uploadPathMock)
-      },
-      { currentItem: currentPathMock, fileUrl: uploadPathMock }
-    )
-  })
-
   it('converts normal files to uppy resources', () => {
     createWrapper(
       () => {
         const fileName = 'filename'
-        const { inputFilesToUppyFiles } = useUploadHelpers()
+        const { inputFilesToUppyFiles } = useUploadHelpers({
+          space: mockDeep<ComputedRef<SpaceResource>>(),
+          currentFolder: computed(() => '')
+        })
         const uppyResources = inputFilesToUppyFiles([{ name: fileName }])
         expect(uppyResources.length).toBe(1)
 
         for (const uppyResource of uppyResources) {
+          // TODO: this would probably need some more checks on props and a proper space mock.
           expect(uppyResource.name).toBe(fileName)
           expect(uppyResource.meta).not.toBeUndefined()
         }
